+++ src/main/kotlin/blackjack/controller/PockerController.kt
@@ -18,26 +18,18 @@ class PockerController {
         OutputView.printInitialState(people)
 
         people.forEach { person ->
-            pockerMachine.addCard(retryOrNot(), person)
+            pockerMachine.addCard(retryOrNot(), person, OutputView::printCardState)
             OutputView.printCardState(person)
         }
         OutputView.printResult(people)
     }
 
     private fun retryOrNot() = { person: Person ->
-        val input = InputView.readRetry(person.name)
-        validateRetryInput(input)
-        person.getScore() < MAXIMUM_SCORE && input == RETRY
-    }
-
-    private fun validateRetryInput(input: String) {
-        require(input in listOf(RETRY, RETRY_NOT)) { RETRY_INPUT_EXCEPTION }
+        val retryOrNot = InputView.retryOrNot(person.name)
+        person.getScore() < MAXIMUM_SCORE && retryOrNot
     }
 
     companion object {
         private const val MAXIMUM_SCORE = 21
-        private const val RETRY = "y"
-        private const val RETRY_NOT = "n"
-        private const val RETRY_INPUT_EXCEPTION = "y 혹은 n을 입력해야만해요."
     }
 }
